+++ TR2Main-VS/specific/texture.h
@@ -38,7 +38,7 @@ int GetTextureSideByHandle(HWR_TEXHANDLE handle);
 
 HRESULT LoadTextureFromFile(LPCTSTR fileName, UINT usage, D3DFORMAT format, D3DPOOL pool, LPDIRECT3DTEXTURE9* pResult);
 void CopyBitmapPalette(RGB888* srcPal, BYTE* srcBitmap, int bitmapSize, RGB888* destPal); // 0x00455990
-BYTE FindNearestPaletteEntry(RGB888* palette, int red, int green, int blue, bool ignoreSysPalette); // 0x00455AD0
+BYTE FindNearestPaletteEntry(RGB888* palette, BYTE red, BYTE green, BYTE blue, bool ignoreSysPalette); // 0x00455AD0
 void SyncSurfacePalettes(void* srcData, int width, int height, int srcPitch, RGB888* srcPalette, void* dstData, int dstPitch, RGB888* dstPalette, bool preserveSysPalette); // 0x00455BA0
 int CreateTexturePalette(RGB888* pal); // 0x00455C50
 int GetFreePaletteIndex(); // 0x00455CE0
